move to common code
