+++ pkg/detectors/alchemy/alchemy_test.go
@@ -6,9 +6,6 @@ package alchemy
 import (
 	"context"
 	"fmt"
-	"io"
-	"net/http"
-	"strings"
 	"testing"
 	"time"
 
@@ -107,7 +104,7 @@ func TestAlchemy_FromChunk(t *testing.T) {
 		},
 		{
 			name: "found, verified but unexpected api surface",
-			s:    Scanner{client: constantStatusHttpClient(404)},
+			s:    Scanner{client: common.ConstantStatusHttpClient(404)},
 			args: args{
 				ctx:    context.Background(),
 				data:   []byte(fmt.Sprintf("You can find a alchemy secret %s within", secret)),
@@ -159,21 +156,6 @@ func BenchmarkFromData(benchmark *testing.B) {
 				}
 			}
 		})
-	}
-}
-
-func constantStatusHttpClient(statusCode int) *http.Client {
-	return &http.Client{
-		Timeout: common.DefaultResponseTimeout,
-		Transport: common.FakeTransport{
-			CreateResponse: func(req *http.Request) (*http.Response, error) {
-				return &http.Response{
-					Request:    req,
-					Body:       io.NopCloser(strings.NewReader("")),
-					StatusCode: statusCode,
-				}, nil
-			},
-		},
 	}
 }
